split app into src/lib/core/*
