+++ src/lib/core/determineTask.ts
@@ -1,0 +1,18 @@
+import { WacLdpTask, parseHttpRequest, TaskType } from '../api/http/HttpParser'
+import uuid from 'uuid/v4'
+import Debug from 'debug'
+const debug = Debug('determineTask')
+
+export async function determineTask (httpReq) {
+  const wacLdpTask: WacLdpTask = await parseHttpRequest(httpReq)
+  // convert ContainerMemberAdd tasks to WriteBlob tasks on the new child
+  if (wacLdpTask.ldpTaskType === TaskType.containerMemberAdd) {
+    debug('converting', wacLdpTask)
+    wacLdpTask.path = wacLdpTask.path.toChild(uuid())
+    wacLdpTask.ldpTaskType = TaskType.blobWrite
+    wacLdpTask.isContainer = false
+    debug('converted', wacLdpTask)
+  }
+  debug('parsed', wacLdpTask)
+  return wacLdpTask
+}
